fix(seed): move mock-utils source to src/ to fix turbo cache and exports
