+++ packages/commons/mock-utils/test/convertToWireMock.test.ts
@@ -2,7 +2,7 @@ import { FernIr } from "@fern-fern/ir-sdk";
 import { IntermediateRepresentation as IRSerializer } from "@fern-fern/ir-sdk/serialization";
 import { readFileSync } from "fs";
 import { join } from "path";
-import { WireMock } from "../index.js";
+import { WireMock } from "../src/index.js";
 
 const loadIr = (irPath: string): FernIr.IntermediateRepresentation => {
     const irJson = JSON.parse(readFileSync(irPath, "utf-8"));
